+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/garmin/watches/tactix/GarminTactix8Coordinator.java
@@ -1,0 +1,18 @@
+package nodomain.freeyourgadget.gadgetbridge.devices.garmin.watches.tactix;
+
+import java.util.regex.Pattern;
+
+import nodomain.freeyourgadget.gadgetbridge.R;
+import nodomain.freeyourgadget.gadgetbridge.devices.garmin.watches.GarminWatchCoordinator;
+
+public class GarminTactix8Coordinator extends GarminWatchCoordinator {
+    @Override
+    protected Pattern getSupportedDeviceName() {
+        return Pattern.compile("^tactix 8( - [\\d+]+mm)?$");
+    }
+
+    @Override
+    public int getDeviceNameResource() {
+        return R.string.devicetype_garmin_tactix_8;
+    }
+}