+++ app/src/test/java/nodomain/freeyourgadget/gadgetbridge/devices/AbstractDeviceCoordinatorTest.java
@@ -98,6 +98,7 @@ public class AbstractDeviceCoordinatorTest extends TestBase {
             put("Venu Sq", DeviceType.GARMIN_VENU_SQ);
             put("fenix 8 - 51mm", DeviceType.GARMIN_FENIX_8); // #4226
             put("quatix 8 - 51mm", DeviceType.GARMIN_QUATIX_8); // #5575
+            put("tactix 8 - 51mm", DeviceType.GARMIN_TACTIX_8); // #5772
             put("Forerunner 265S", DeviceType.GARMIN_FORERUNNER_265S);
             put("Forerunner 955", DeviceType.GARMIN_FORERUNNER_955); // #4124
             put("Enduro", DeviceType.GARMIN_ENDURO); // #5347